refactor: move the tap option parsers into lib/tap/utils
Review feedback. parseIndex and parsePositiveInt parse raw commander option
strings — nothing about them is the AUT frame, and only the commands import
them. They move to lib/tap/utils alongside their test, so aut/frame is left
with the frame resolution and the lifecycle gate it is named for.
